+++ changelog.md
@@ -5,6 +5,7 @@
 * Rebalanced novice farmer and fletcher trades
 * The regenerating absorption effect now increases regen speed
 * Lowered Fire Aspect seconds on fire (4 -> 3 per level) and the seconds on fire now scale with player attack cooldown
+* Heavily lowered Knockback strength (1 -> 0.4 per level)
 
 ### Fixes
 * Luck enchantment being harder to obtain compared to Silk Touch